+++ upyog/ml/imports.py
@@ -1,11 +1,13 @@
 from upyog.imports import *
 from torch.utils.data import Dataset, DataLoader
 from torch import Tensor
+from torchvision.io.image import ImageReadMode
 
 import torch
 import torch.nn as nn
 import torch.nn.functional as F
 import torch.optim as optim
+import torchvision.transforms as Tfms
 import torchvision.transforms.functional as TF
 
 import matplotlib as mpl